Unified ALLForSaleProperty Listing with CQRS Pattern :
This feature merges housing, land, and commercial property listings into a single unified ViewModel called ForSalePropertyForListingViewModel, which includes shared and type-specific fields. The data is retrieved from separate tables using LINQ queries in the repository and combined using Concat() with agent details. Following the CQRS pattern, we defined a Query, Result  and Handler to process and return the data in a clean, maintainable structure.